# backend/adapter/controller/handler/task.go
package handler

import (
	"backend/adapter/controller/presenter"
	"backend/api"
	"backend/entity"
	"backend/pkg/logger"
	"backend/usecase"
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
)

type ITaskHandler interface {
	CreateTask(c *gin.Context)
	GetTaskById(c *gin.Context, id int)
	GetAllTasks(c *gin.Context)
	UpdateTaskById(c *gin.Context, id int)
	DeleteTaskById(c *gin.Context, id int)
}

type taskHandler struct {
	tu usecase.ITaskUsecase
}

func NewTaskHandler(tu usecase.ITaskUsecase) ITaskHandler {
	return &taskHandler{tu: tu}
}

func taskToResponse(task *entity.Task) presenter.TaskResponse {
	return presenter.TaskResponse{
		ApiVersion: api.Version,
		Data: presenter.Task{
			Kind: "task",
			Id:   int(task.ID),
			Name: task.Name,
			Status: presenter.Status{
				Id:   (*int)(&task.Status.ID),
				Name: presenter.StatusName(task.Status.Name),
			},
		},
	}
}

func tasksToResponse(tasks *[]entity.Task) presenter.TasksResponse {
	data := make([]presenter.Task, len(*tasks))
	for i, task := range *tasks {
		data[i] = presenter.Task{
			Kind: "task",
			Id:   int(task.ID),
			Name: task.Name,
			Status: presenter.Status{
				Id:   (*int)(&task.Status.ID),
				Name: presenter.StatusName(task.Status.Name),
			},
		}
	}
	return presenter.TasksResponse{
		ApiVersion: api.Version,
		Data:       data,
	}
}

func getUserIDFromContext(c *gin.Context) (entity.UserID, error) {
	userID, exists := c.Get("user_id")
	if !exists {
		logger.Warn("user_id not found in context")
		return 0, fmt.Errorf("user_id not found in context")
	}

	userIDFloat, ok := userID.(float64)
	if !ok {
		logger.Warn(fmt.Sprintf("user_id has invalid type: %T", userID))
		return 0, fmt.Errorf("invalid user_id type")
	}

	return entity.UserID(userIDFloat), nil
}

func (th *taskHandler) CreateTask(c *gin.Context) {
	var requestBody presenter.CreateTaskRequestBody
	if err := c.ShouldBindJSON(&requestBody); err != nil {
		logger.Warn(err.Error())
		c.JSON(presenter.NewErrorResponse(http.StatusBadRequest, err.Error()))
		return
	}

	status, err := entity.NewStatus(string(requestBody.Status.Name))
	if err != nil {
		logger.Warn(err.Error())
		c.JSON(presenter.NewErrorResponse(http.StatusBadRequest, err.Error()))
		return
	}

	userID, err := getUserIDFromContext(c)
	if err != nil {
		logger.Warn(err.Error())
		c.JSON(presenter.NewErrorResponse(http.StatusUnauthorized, err.Error()))
		return
	}

	task := &entity.Task{
		Name:   requestBody.Name,
		Status: *status,
		UserID: userID,
	}

	createdTask, err := th.tu.Create(task)
	if err != nil {
		logger.Error(err.Error())
		c.JSON(presenter.NewErrorResponse(http.StatusInternalServerError, err.Error()))
		return
	}
	c.JSON(http.StatusCreated, taskToResponse(createdTask))
}

func (th *taskHandler) GetTaskById(c *gin.Context, id int) {
	userID, err := getUserIDFromContext(c)
	if err != nil {
		logger.Warn(err.Error())
		c.JSON(presenter.NewErrorResponse(http.StatusUnauthorized, err.Error()))
		return
	}

	taskID := entity.TaskID(id)

	task, err := th.tu.Get(taskID, userID)
	if err != nil {
		logger.Error(err.Error())
		c.JSON(presenter.NewErrorResponse(http.StatusInternalServerError, err.Error()))
		return
	}
	c.JSON(http.StatusOK, taskToResponse(task))
}

func (th *taskHandler) GetAllTasks(c *gin.Context) {
	userID, err := getUserIDFromContext(c)
	if err != nil {
		logger.Warn(err.Error())
		c.JSON(presenter.NewErrorResponse(http.StatusUnauthorized, err.Error()))
		return
	}

	tasks, err := th.tu.GetAll(userID)
	if err != nil {
		logger.Error(err.Error())
		c.JSON(presenter.NewErrorResponse(http.StatusInternalServerError, err.Error()))
		return
	}
	c.JSON(http.StatusOK, tasksToResponse(tasks))
}

func (th *taskHandler) UpdateTaskById(c *gin.Context, id int) {
	var requestBody presenter.UpdateTaskRequestBody
	if err := c.ShouldBindJSON(&requestBody); err != nil {
		logger.Warn(err.Error())
		c.JSON(presenter.NewErrorResponse(http.StatusBadRequest, err.Error()))
		return
	}

	status, err := entity.NewStatus(string(requestBody.Status.Name))
	if err != nil {
		logger.Warn(err.Error())
		c.JSON(presenter.NewErrorResponse(http.StatusBadRequest, err.Error()))
		return
	}

	userID, err := getUserIDFromContext(c)
	if err != nil {
		logger.Warn(err.Error())
		c.JSON(presenter.NewErrorResponse(http.StatusUnauthorized, err.Error()))
		return
	}

	taskID := entity.TaskID(id)

	task := &entity.Task{
		ID:     taskID,
		Name:   requestBody.Name,
		Status: *status,
		UserID: userID,
	}

	updatedTask, err := th.tu.Save(task)
	if err != nil {
		logger.Error(err.Error())
		c.JSON(presenter.NewErrorResponse(http.StatusInternalServerError, err.Error()))
		return
	}
	c.JSON(http.StatusOK, taskToResponse(updatedTask))
}

func (th *taskHandler) DeleteTaskById(c *gin.Context, id int) {
	userID, err := getUserIDFromContext(c)
	if err != nil {
		c.JSON(presenter.NewErrorResponse(http.StatusUnauthorized, err.Error()))
		return
	}

	taskID := entity.TaskID(id)

	if err := th.tu.Delete(taskID, userID); err != nil {
		c.JSON(presenter.NewErrorResponse(http.StatusInternalServerError, err.Error()))
		return
	}
	c.Status(http.StatusNoContent)
}