# backend/adapter/controller/handler/user.go
package handler

import (
	"backend/adapter/controller/presenter"
	"backend/api"
	"backend/entity"
	"backend/pkg/cookie"
	"backend/pkg/logger"
	"backend/usecase"
	"net/http"

	"github.com/gin-gonic/gin"
)

type IUserHandler interface {
	PostSignUp(c *gin.Context)
	PostLogin(c *gin.Context)
	PostLogout(c *gin.Context)
}

type userHandler struct {
	uu usecase.IUserUsecase
}

func NewUserHandler(uu usecase.IUserUsecase) IUserHandler {
	return &userHandler{uu: uu}
}

func (uh *userHandler) PostSignUp(c *gin.Context) {
	var requestBody presenter.SignUpRequestBody
	if err := c.ShouldBindJSON(&requestBody); err != nil {
		logger.Warn(err.Error())
		c.JSON(presenter.NewErrorResponse(http.StatusBadRequest, err.Error()))
		return
	}

	user := &entity.User{
		Email:    requestBody.User.Email,
		Password: *requestBody.User.Password,
	}

	// 平文パスワードを保存（Login用）
	plainPassword := *requestBody.User.Password

	createdUser, err := uh.uu.SignUp(user)
	if err != nil {
		logger.Error(err.Error())
		c.JSON(presenter.NewErrorResponse(http.StatusInternalServerError, err.Error()))
		return
	}

	// 自動ログインのためにLoginユースケースを呼び出す
	// 平文パスワードを持つUserオブジェクトを作成
	loginUser := &entity.User{
		Email:    createdUser.Email,
		Password: plainPassword,
	}
	tokenString, err := uh.uu.Login(loginUser)
	if err != nil {
		logger.Error((err.Error()))
		c.JSON(presenter.NewErrorResponse(http.StatusInternalServerError, err.Error()))
		return
	}

	sameSite, secure, domain := cookie.GetCookieConfig()

	http.SetCookie(c.Writer, &http.Cookie{
		Name:     "token",
		Value:    tokenString,
		MaxAge:   24 * 60 * 60,
		Path:     "/",
		Domain:   domain,
		Secure:   secure,
		HttpOnly: true,
		SameSite: sameSite,
	})

	userID := int(createdUser.ID)

	c.JSON(http.StatusCreated, presenter.SignUpResponse{
		ApiVersion: api.Version,
		Data: presenter.User{
			Kind:  "user",
			Id:    &userID,
			Email: createdUser.Email,
		},
	})
}

func (uh *userHandler) PostLogin(c *gin.Context) {
	var requestBody presenter.LoginRequestBody
	if err := c.ShouldBindJSON(&requestBody); err != nil {
		logger.Warn(err.Error())
		c.JSON(presenter.NewErrorResponse(http.StatusBadRequest, err.Error()))
		return
	}

	user := &entity.User{
		Email:    requestBody.User.Email,
		Password: *requestBody.User.Password,
	}

	tokenString, err := uh.uu.Login(user)
	if err != nil {
		logger.Error((err.Error()))
		c.JSON(presenter.NewErrorResponse(http.StatusInternalServerError, err.Error()))
		return
	}

	sameSite, secure, domain := cookie.GetCookieConfig()

	http.SetCookie(c.Writer, &http.Cookie{
		Name:     "token",
		Value:    tokenString,
		MaxAge:   24 * 60 * 60,
		Path:     "/",
		Domain:   domain,
		Secure:   secure,
		HttpOnly: true,
		SameSite: sameSite,
	})
	c.Status(http.StatusOK)
}

func (uh *userHandler) PostLogout(c *gin.Context) {
	sameSite, secure, domain := cookie.GetCookieConfig()

	http.SetCookie(c.Writer, &http.Cookie{
		Name:     "token",
		Value:    "",
		MaxAge:   -1,
		Path:     "/",
		Domain:   domain,
		Secure:   secure,
		HttpOnly: true,
		SameSite: sameSite,
	})
	c.Status(http.StatusOK)
}